Rename module `Internal.Invariant`.
Rename module `Internal.Invariant` to `Internal.Data.Invariant`.
